CUDASetup: added 64-bit build check, fix multiple-include-dir bug
CUDA's support for 32-bit builds is mostly removed, and the few
parts that remain are deprecated. So, I just went ahead and added
a check to CUDASetup to throw a fatal error if we're doing a 32-bit
build with CUDA.

CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES currently has only one
directory in it, but there's no guarantee that will always be the
case. So, added fixes so that the code will still work correctly
if this var contains multiple dirs.

Verified the above fix by manually hacking the file to put additional
directories in CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES. Works.